+++ changelog.md
@@ -2,6 +2,9 @@
 
 **pycite 0.1.2**
 
+
+* Now supporting JSTOR as a paper source. See [#15](https://github.com/Nelson-Gon/pycite/issues/15).
+
 * Fixed issues with misplaced "&" in citations. See [#8](https://github.com/Nelson-Gon/pycite/issues/8).
 
 * Fixed issues with mix up of science direct volumes, years, pages. See [#7](https://github.com/Nelson-Gon/pycite/issues/7).